Flytter ClientConfigurationPropertiesMatcher bønnen ut i egen klasse.
